chore: :construction: refactored with auth controller
refactored user controller and auth controller
